initial commit. ERROR

+++ include/mfigp/ext/block.h
@@ -144,7 +144,7 @@ public:
 			if (last && last->next) last = last -> next;
 			else
 			{
-				block *next = (block *) new char [sizeof(block) + (block_size-1)*sizeof(Type)];
+				block *next = (block *) new char [sizeof(block) + size_t(block_size-1)*sizeof(Type)];
 				if (!next) { if (error_function) (*error_function)("Not enough memory!"); exit(1); }
 				if (last) last -> next = next;
 				else first = next;
@@ -270,7 +270,7 @@ public:
 		if (!first_free)
 		{
 			block *next = first;
-			first = (block *) new char [sizeof(block) + (block_size-1)*sizeof(block_item)];
+			first = (block *) new char [sizeof(block) + size_t(block_size-1)*sizeof(block_item)];
 			if (!first) { if (error_function) (*error_function)("Not enough memory!"); exit(1); }
 			first_free = & (first -> data[0] );
 			for (item=first_free; item<first_free+block_size-1; item++)
